requirements.txt
streamlit
google-generativeai
PyPDF2
docx2txt
requests
beautifulsoup4
selenium
webdriver-manager
spacy
langchain
langchain-google-genai
faiss-cpu